Rename RegistryClient to ExecutorClient and add ActionExecutionError
